webrtc-extensions/RTCRtpReceiver-playoutDelayHint.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>RTCRtpReceiver-playoutDelayHint</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<body>
<script>
'use strict'

// Test is based on the following draft:
// https://henbos.github.io/webrtc-timing


test(t => {
  const pc = new RTCPeerConnection();
  t.add_cleanup(() => pc.close());
  const {receiver} = pc.addTransceiver('audio', {direction:'recvonly'});
  assert_equals(receiver.playoutDelayHint, null,
      'audio playoutDelayHint is null by default');
});

test(t => {
  const pc = new RTCPeerConnection();
  t.add_cleanup(() => pc.close());
  const {receiver} = pc.addTransceiver('audio', {direction:'recvonly'});
  receiver.playoutDelayHint = 0.5;
  assert_equals(receiver.playoutDelayHint, 0.5,
      'audio playoutDelayHint saves correct value');
});

test(t => {
  const pc = new RTCPeerConnection();
  t.add_cleanup(() => pc.close());
  const {receiver} = pc.addTransceiver('audio', {direction:'recvonly'});
  receiver.playoutDelayHint = 20.5;
  assert_equals(receiver.playoutDelayHint, 20.5,
      'audio playoutDelayHint saves large value');
  pc.close();
});

test(t => {
  const pc = new RTCPeerConnection();
  t.add_cleanup(() => pc.close());
  const {receiver} = pc.addTransceiver('audio', {direction:'recvonly'});
  receiver.playoutDelayHint = 0.7
  assert_throws('InvalidAccessError', () => {
    receiver.playoutDelayHint = -0.5;
  }, 'audio playoutDelayHint accepts only positive values and 0.0');
  assert_equals(receiver.playoutDelayHint, 0.7,
      'audio playoutDelayHint returns last valid value');
});

test(t => {
  const pc = new RTCPeerConnection();
  t.add_cleanup(() => pc.close());
  const {receiver} = pc.addTransceiver('audio', {direction:'recvonly'});
  receiver.playoutDelayHint = 0.0;
  assert_equals(receiver.playoutDelayHint, 0.0,
      'audio playoutDelayHint allows zero value');
  pc.close();
});

test(t => {
  const pc = new RTCPeerConnection();
  t.add_cleanup(() => pc.close());
  const {receiver} = pc.addTransceiver('audio', {direction:'recvonly'});
  receiver.playoutDelayHint = 0.5;
  receiver.playoutDelayHint = null;
  assert_equals(receiver.playoutDelayHint, null,
      'audio playoutDelayHint allows to reset value to null');
  pc.close();
});

test(t => {
  const pc = new RTCPeerConnection();
  t.add_cleanup(() => pc.close());
  const {receiver} = pc.addTransceiver('video', {direction:'recvonly'});
  assert_equals(receiver.playoutDelayHint, null,
      'video playoutDelayHint is null by default');
  pc.close();
});

test(t => {
  const pc = new RTCPeerConnection();
  t.add_cleanup(() => pc.close());
  const {receiver} = pc.addTransceiver('video', {direction:'recvonly'});
  receiver.playoutDelayHint = 0.5;
  assert_equals(receiver.playoutDelayHint, 0.5,
      'video playoutDelayHint saves correct value');
  pc.close();
});

test(t => {
  const pc = new RTCPeerConnection();
  t.add_cleanup(() => pc.close());
  const {receiver} = pc.addTransceiver('video', {direction:'recvonly'});
  receiver.playoutDelayHint = 20.5;
  assert_equals(receiver.playoutDelayHint, 20.5,
      'video playoutDelayHint saves large value');
  pc.close();
});

test(t => {
  const pc = new RTCPeerConnection();
  t.add_cleanup(() => pc.close());
  const {receiver} = pc.addTransceiver('video', {direction:'recvonly'});
  receiver.playoutDelayHint = 0.7
  assert_throws('InvalidAccessError', () => {
    receiver.playoutDelayHint = -0.5;
  }, 'video playoutDelayHint accepts only positive values and 0.0');
  assert_equals(receiver.playoutDelayHint, 0.7,
      'video playoutDelayHint returns last valid value');
  pc.close();
});

test(t => {
  const pc = new RTCPeerConnection();
  t.add_cleanup(() => pc.close());
  const {receiver} = pc.addTransceiver('video', {direction:'recvonly'});
  receiver.playoutDelayHint = 0.0;
  assert_equals(receiver.playoutDelayHint, 0.0,
      'video playoutDelayHint allows zero value');
  pc.close();
});

test(t => {
  const pc = new RTCPeerConnection();
  t.add_cleanup(() => pc.close());
  const {receiver} = pc.addTransceiver('video', {direction:'recvonly'});
  receiver.playoutDelayHint = 0.5;
  receiver.playoutDelayHint = null;
  assert_equals(receiver.playoutDelayHint, null,
      'video playoutDelayHint allows to reset value to null');
  pc.close();
});
</script>
</body>
